Extract more inter-dependent modules
