+++ rs/assets/shaders/terrain.mat.ron.meta
@@ -1,7 +1,7 @@
 (
     meta_format_version: "1.0",
     asset: Load(
-        loader: "sond_has_engine::util::RonReflectAssetLoader<bevy_pbr::extended_material::ExtendedMaterial<bevy_pbr::pbr_material::StandardMaterial, sond_has::mats::fog::DistanceDither>>",
+        loader: "sond_has_engine::util::RonReflectAssetLoader<bevy_pbr::extended_material::ExtendedMaterial<bevy_pbr::pbr_material::StandardMaterial, sond_has_engine::mats::fog::DistanceDither>>",
         settings: (),
     ),
 )
