+++ rs/engine/Cargo.toml
@@ -55,3 +55,8 @@ colored = { version = "2", optional = true }
 
 [target.'cfg(not(target_arch="wasm32"))'.dependencies]
 bevy_dylib = { workspace = true }
+
+[lints.clippy]
+needless_update = "allow" # `..default()` used a lot to allow quick adding/removing of fields
+too_many_arguments = "allow" # Systems often need lots of arguments
+type_complexity = "allow" # Queries seem complex to Clippy but are usually easily read
