+++ rs/engine/src/nav/alg.rs
