+++ rs/engine/src/nav/heightmap.rs
